fix: styled-componentsを別ファイルに切り出し
